Add the initial overwrite of the layout and page
This adds a custom header and footer to the page along with the images
needed by the footer.  The logos were taken from the appropriate
companies' websites according to the social media use.  This is
primarily for my website to keep the contact links consistent.  We also
place the side bar on the left by default.