Adjust the margins to show the sidebar
The default sidebar slides all the way to the left.  I don't want that.
That just means we need a bit more CSS to left align the sidebar.  While
we're at it, we left align the text instead of justify it because the
spacing is so tight.  The only other thing we need to do is wrap the
document block in a container.  We don't want to overwrite the layout
block because that messes with the side bar.  This gets the formatting
more in line with the original source.